fix(monetary): correcciones finales del modelo integer (ADR-018)
CORRECCIONES APLICADAS:

1. CashierPayBillTest.php:
   - Eliminados TODOS los casts (float) del archivo
   - Problema: (float) $responseData['remaining_amount'] convertía
     0 (int) a 0.0 (float), y ->toBe(0) con === fallaba
   - Solución: eliminar casts, los valores ya son integer

2. CashCount.php (cambios previos del usuario):
   - 4 campos decimal:2 → integer (cash_amount, card_amount,
     transfer_amount, other_amount)
   - recalculateFromDenominations(): float → int
   - discrepancyPercentage(): bug del (int) round() eliminado
   - Diferencia se calcula como resta directa de enteros

3. Denomination.php (cambios previos del usuario):
   - calculateTotal(): float → int
   - Retorna suma de denominaciones CLP como entero

SEPARACIÓN MANTENIDA:
💰 Dinero CLP → integer
📊 Porcentajes (discrepancyPercentage) → float con 2 decimales
⚖️  Cantidades físicas → float (sin cambios)

RESULTADO:
✅ Todos los tests de Cashier pasan
✅ Modelo monetario 100% integer end-to-end

+++ app/Modules/Cashier/Domain/ValueObjects/Denomination.php
@@ -89,9 +89,9 @@ class Denomination
      * Calcula el monto total a partir de un array de conteos.
      * Formato esperado: ['20000' => 5, '10000' => 3, ...]
      */
-    public static function calculateTotal(array $counts): float
+    public static function calculateTotal(array $counts): int
     {
-        $total = 0.0;
+        $total = 0;
         foreach ($counts as $denomination => $quantity) {
             if (self::isValid((int) $denomination) && is_numeric($quantity)) {
                 $total += ((int) $denomination) * ((int) $quantity);
